fix: fixed breadcrumb routes and organized tsx & style files in the same folder
