+++ src/constant/dynamic-breadcrumb-title.ts
@@ -1,0 +1,8 @@
+export const DYNAMIC_BREADCRUMB_TITLES: Record<string, string | null> = {
+  '/dashboard': 'Dashboard',
+  '/inventory': 'Inventory',
+  '/identity-management': 'IAM',
+  '/inventory/[id]': 'Inventory',
+  '/profile': 'Profile',
+  '/activity-log': 'Activity log',
+};
